books

+++ books/controller/delete.php
@@ -14,8 +14,10 @@ class Delete extends Select
 
 	public function Delete($name, $folder){
 		$filename = "../files/" . $folder . $name;
-		if(file_exists($filename)){
+		if ($name) {
+			if(file_exists($filename)){
 			unlink($filename);
+			}
 		}
 	}
 	public function DelImg(){
@@ -27,7 +29,21 @@ class Delete extends Select
 			$gel_img_db = parent::update("UPDATE books SET book_img = :img_name WHERE book_id = '$id'", $param);
 		}
 	}
+
+	public function DelPost(){
+		if ($_GET['bookdel']) {
+			$id = $_GET['bookdel'];
+			$img_name = parent::selectAll("SELECT book_img FROM books WHERE book_id  = '$id'");
+			$del_file = $this->Delete($img_name[0]['book_img'], "img/");
+			$file_name = parent::selectAll("SELECT book_file FROM books WHERE book_id  = '$id'");
+			$del_file = $this->Delete($file_name[0]['book_file'],"");
+			$del_post = parent::DeleteItem("DELETE FROM books WHERE book_id  = '$id'");
+			$del_rate = parent::DeleteItem("DELETE FROM rate WHERE book_id  = '$id'");
+			header("Location: ../view/cabinet.php");
+		}
+	}
 }
 
 $del = new Delete();
 $del_img = $del->DelImg();
+$del_post = $del->DelPost();
